refactor(markdown): Decompose Markdown component for improved modularity
Splits the  file into separate modules for block renderers (CodeBlock, VideoBlock, etc.), utility functions, and ErrorBoundary. Introduces a barrel file for block components and adds documentation.
