+++ web/app/components/base/markdown-blocks/index.ts
@@ -1,0 +1,18 @@
+/**
+ * @fileoverview Barrel file for all markdown block components.
+ * This allows for cleaner imports in other parts of the application.
+ */
+
+export { default as AudioBlock } from './audio-block';
+export { default as CodeBlock } from './code-block';
+export { default as Img } from './img';
+export { default as Link } from './link';
+export { default as Paragraph } from './paragraph';
+export { default as PreCode } from './pre-code';
+export { default as ScriptBlock } from './script-block';
+export { default as VideoBlock } from './video-block';
+
+// Assuming these are also standalone components in this directory intended for Markdown rendering
+export { default as MarkdownButton } from './button';
+export { default as MarkdownForm } from './form';
+export { default as ThinkBlock } from './think-block';
