+++ web/app/components/base/markdown/index.tsx
@@ -5,126 +5,29 @@ import RemarkBreaks from 'remark-breaks'
 import RehypeKatex from 'rehype-katex'
 import RemarkGfm from 'remark-gfm'
 import RehypeRaw from 'rehype-raw'
-import { Component, memo, useRef } from 'react'
 import { flow } from 'lodash-es'
-import ImageGallery from '@/app/components/base/image-gallery'
-import { useChatContext } from '@/app/components/base/chat/chat/context'
-import VideoGallery from '@/app/components/base/video-gallery'
-import AudioGallery from '@/app/components/base/audio-gallery'
-import MarkdownButton from '@/app/components/base/markdown-blocks/button'
-import MarkdownForm from '@/app/components/base/markdown-blocks/form'
-import ThinkBlock from '@/app/components/base/markdown-blocks/think-block'
 import cn from '@/utils/classnames'
-import CodeBlock from '@/app/components/base/markdown-blocks/code-block'
+import { preprocessLaTeX, preprocessThinkTag } from './markdown-utils'
+import {
+  CodeBlock,
+  Paragraph,
+  VideoBlock,
+  AudioBlock,
+  ScriptBlock,
+  Img,
+  Link,
+  MarkdownButton,
+  MarkdownForm,
+  ThinkBlock
+} from '@/app/components/base/markdown-blocks'
 
-const preprocessLaTeX = (content: string) => {
-  if (typeof content !== 'string')
-    return content
-
-  const codeBlockRegex = /```[\s\S]*?```/g
-  const codeBlocks = content.match(codeBlockRegex) || []
-  let processedContent = content.replace(codeBlockRegex, 'CODE_BLOCK_PLACEHOLDER')
-
-  processedContent = flow([
-    (str: string) => str.replace(/\\\[(.*?)\\\]/g, (_, equation) => `$$${equation}$$`),
-    (str: string) => str.replace(/\\\[([\s\S]*?)\\\]/g, (_, equation) => `$$${equation}$$`),
-    (str: string) => str.replace(/\\\((.*?)\\\)/g, (_, equation) => `$$${equation}$$`),
-    (str: string) => str.replace(/(^|[^\\])\$(.+?)\$/g, (_, prefix, equation) => `${prefix}$${equation}$`),
-  ])(processedContent)
-
-  codeBlocks.forEach((block) => {
-    processedContent = processedContent.replace('CODE_BLOCK_PLACEHOLDER', block)
-  })
-
-  return processedContent
-}
-
-const preprocessThinkTag = (content: string) => {
-  const thinkOpenTagRegex = /<think>\n/g
-  const thinkCloseTagRegex = /\n<\/think>/g
-  return flow([
-    (str: string) => str.replace(thinkOpenTagRegex, '<details data-think=true>\n'),
-    (str: string) => str.replace(thinkCloseTagRegex, '\n[ENDTHINKFLAG]</details>'),
-  ])(content)
-}
-
-export function PreCode(props: { children: any }) {
-  const ref = useRef<HTMLPreElement>(null)
-
-  return (
-    <pre ref={ref}>
-      <span
-        className="copy-code-button"
-      ></span>
-      {props.children}
-    </pre>
-  )
-}
-
-const VideoBlock: any = memo(({ node }: any) => {
-  const srcs = node.children.filter((child: any) => 'properties' in child).map((child: any) => (child as any).properties.src)
-  if (srcs.length === 0) {
-    const src = node.properties?.src
-    if (src)
-      return <VideoGallery key={src} srcs={[src]} />
-    return null
-  }
-  return <VideoGallery key={srcs.join()} srcs={srcs} />
-})
-VideoBlock.displayName = 'VideoBlock'
-
-const AudioBlock: any = memo(({ node }: any) => {
-  const srcs = node.children.filter((child: any) => 'properties' in child).map((child: any) => (child as any).properties.src)
-  if (srcs.length === 0) {
-    const src = node.properties?.src
-    if (src)
-      return <AudioGallery key={src} srcs={[src]} />
-    return null
-  }
-  return <AudioGallery key={srcs.join()} srcs={srcs} />
-})
-AudioBlock.displayName = 'AudioBlock'
-
-const ScriptBlock = memo(({ node }: any) => {
-  const scriptContent = node.children[0]?.value || ''
-  return `<script>${scriptContent}</script>`
-})
-ScriptBlock.displayName = 'ScriptBlock'
-
-const Paragraph = (paragraph: any) => {
-  const { node }: any = paragraph
-  const children_node = node.children
-  if (children_node && children_node[0] && 'tagName' in children_node[0] && children_node[0].tagName === 'img') {
-    return (
-      <div className="markdown-img-wrapper">
-        <ImageGallery srcs={[children_node[0].properties.src]} />
-        {
-          Array.isArray(paragraph.children) && paragraph.children.length > 1 && (
-            <div className="mt-2">{paragraph.children.slice(1)}</div>
-          )
-        }
-      </div>
-    )
-  }
-  return <p>{paragraph.children}</p>
-}
-
-const Img = ({ src }: any) => {
-  return <div className="markdown-img-wrapper"><ImageGallery srcs={[src]} /></div>
-}
-
-const Link = ({ node, children, ...props }: any) => {
-  if (node.properties?.href && node.properties.href?.toString().startsWith('abbr')) {
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    const { onSend } = useChatContext()
-    const hidden_text = decodeURIComponent(node.properties.href.toString().split('abbr:')[1])
-
-    return <abbr className="cursor-pointer underline !decoration-primary-700 decoration-dashed" onClick={() => onSend?.(hidden_text)} title={node.children[0]?.value || ''}>{node.children[0]?.value || ''}</abbr>
-  }
-  else {
-    return <a {...props} target="_blank" className="cursor-pointer underline !decoration-primary-700 decoration-dashed">{children || 'Download'}</a>
-  }
-}
+/**
+ * @fileoverview Main Markdown rendering component.
+ * This file was refactored to extract individual block renderers and utility functions
+ * into separate modules for better organization and maintainability as of [Date of refactor].
+ * Further refactoring candidates (custom block components not fitting general categories)
+ * are noted in their respective files if applicable.
+ */
 
 export function Markdown(props: { content: string; className?: string; customDisallowedElements?: string[] }) {
   const latexContent = flow([
@@ -145,7 +48,7 @@ export function Markdown(props: { content: string; className?: string; customDis
           RehypeRaw as any,
           // The Rehype plug-in is used to remove the ref attribute of an element
           () => {
-            return (tree) => {
+            return (tree: any) => {
               const iterate = (node: any) => {
                 if (node.type === 'element' && node.properties?.ref)
                   delete node.properties.ref
@@ -181,30 +84,4 @@ export function Markdown(props: { content: string; className?: string; customDis
       </ReactMarkdown>
     </div>
   )
-}
-
-// **Add an ECharts runtime error handler
-// Avoid error #7832 (Crash when ECharts accesses undefined objects)
-// This can happen when a component attempts to access an undefined object that references an unregistered map, causing the program to crash.
-
-export default class ErrorBoundary extends Component {
-  constructor(props: any) {
-    super(props)
-    this.state = { hasError: false }
-  }
-
-  componentDidCatch(error: any, errorInfo: any) {
-    this.setState({ hasError: true })
-    console.error(error, errorInfo)
-  }
-
-  render() {
-    // eslint-disable-next-line ts/ban-ts-comment
-    // @ts-expect-error
-    if (this.state.hasError)
-      return <div>Oops! An error occurred. This could be due to an ECharts runtime error or invalid SVG content. <br />(see the browser console for more information)</div>
-    // eslint-disable-next-line ts/ban-ts-comment
-    // @ts-expect-error
-    return this.props.children
-  }
 }
